Moved Language components to component dir and updated paths
